Update Films Page Added

+++ front/scripts/deleteFilm.js
@@ -2,6 +2,10 @@ const axios = require("axios")
 
 document.getElementById("button-delete").addEventListener("click", () => {
     const filmTitle = document.getElementById("film-title")
+    if (!filmTitle.value) {
+        alert("Please enter a film title");
+        return;
+    }
     axios.delete(`http://localhost:3000/movies/delete-film-byTitle/${filmTitle.value}`)
     .then(res => {
         alert(res.data.message);
